ap - factor out common code in web tests to webtestcase
